refactor: move out logger into it's own module
